requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
psycopg[binary]==3.2.1
psycopg-pool==3.2.1
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.2.0
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
redis==5.0.1
